Second Commit
Added the Models classes for the HotelDB

Created the HotelDbContext

Ran the Initial Migration without the DbSet Enitity configuration which resulted in a database without tables

Updated the database

Added the DbSet Entities into the HotelDbContext as well as defined the decimal column types.

Added the second migration "AddedTables" to the project. Then updated the database.

The tables were created with no errors from the NPM Console and so I am good.