Third Commit
Edited the Guest and Reservation and changed the DateTime time types to string. I will have to figure out a better way to tackle this issue.

Also created a new Migration for that.

Added Dummy DB values into the Guest table.

Added a Scaffolded GuestsController class in the Controllers class.

Ran the project and the API had a working GET method so I am pleased